requirements.txt
python-escpos
python-telegram-bot
sqlalchemy
